--- boa/contracts/vyper/__init__.py
@@ -1,2 +1,0 @@
-# id used internally for method id name
-_METHOD_ID_VAR = "_calldata_method_id"
