+++ boa/contracts/vyper/vyper_contract.py
@@ -32,14 +32,16 @@ from vyper.semantics.analysis.data_positions import set_data_positions
 from vyper.semantics.types import AddressT, HashMapT, TupleT
 from vyper.utils import method_id
 
+from boa.boa_error import BoaError
 from boa.contracts.evm_contract import BaseEVMContract
-from boa.contracts.vyper import _METHOD_ID_VAR
+from boa.contracts.stack_trace import StackTrace, _handle_child_trace
 from boa.contracts.vyper.ast_utils import (
     ast_map_of,
     get_fn_ancestor_from_node,
     reason_at,
 )
 from boa.contracts.vyper.compiler_utils import (
+    _METHOD_ID_VAR,
     anchor_compiler_settings,
     compile_vyper_function,
     generate_bytecode_for_arbitrary_stmt,
@@ -54,7 +56,6 @@ from boa.contracts.vyper.ir_executor import executor_from_ir
 from boa.environment import Address, Env
 from boa.profiling import LineProfile, cache_gas_used_for_computation
 from boa.util.abi import abi_decode, abi_encode
-from boa.util.exceptions import BoaError, StackTrace, _handle_child_trace
 from boa.util.lrudict import lrudict
 from boa.vm.gas_meters import ProfilingGasMeter
 from boa.vm.utils import to_bytes, to_int
@@ -113,7 +114,7 @@ class VyperDeployer:
 
 
 # a few lines of shared code between VyperBlueprint and VyperContract
-class _BaseContract(BaseEVMContract):
+class _BaseVyperContract(BaseEVMContract):
     def __init__(
         self,
         compiler_data: CompilerData,
@@ -130,7 +131,7 @@ class _BaseContract(BaseEVMContract):
 # create a blueprint for use with `create_from_blueprint`.
 # uses a ERC5202 preamble, when calling `create_from_blueprint` will
 # need to use `code_offset=3`
-class VyperBlueprint(_BaseContract):
+class VyperBlueprint(_BaseVyperContract):
     def __init__(
         self,
         compiler_data,
@@ -438,7 +439,7 @@ class ImmutablesModel:
         return repr(self.dump())
 
 
-class VyperContract(_BaseContract):
+class VyperContract(_BaseVyperContract):
     def __init__(
         self,
         compiler_data,
