+++ boa/interpret.py
@@ -7,13 +7,16 @@ import vyper
 from vyper.cli.vyper_compile import get_interface_codes
 from vyper.compiler.phases import CompilerData
 
-from boa.contracts.abi.contract import ABIContractFactory
+from boa.contracts.abi.abi_contract import ABIContractFactory
 from boa.contracts.vyper.compiler_utils import anchor_compiler_settings
-from boa.contracts.vyper.contract import VyperBlueprint, VyperContract, VyperDeployer
+from boa.contracts.vyper.vyper_contract import (
+    VyperBlueprint,
+    VyperContract,
+    VyperDeployer,
+)
 from boa.environment import Address
 from boa.explorer import fetch_abi_from_etherscan
 from boa.util.disk_cache import DiskCache
-from boa.util.exceptions import BoaError
 
 _Contract = Union[VyperContract, VyperBlueprint]
 
@@ -122,4 +125,4 @@ def from_etherscan(
     return ABIContractFactory.from_abi_dict(abi, name=name).at(addr)
 
 
-__all__ = ["BoaError"]
+__all__ = []
